refactor(components): :recycle: extract PostsNavigation with stories
